💄 refactor: simplify carousel components for homepage
